moved filter components into subdirectory
