+++ vis/js/components/filtersort/SortDropdown.js
@@ -1,13 +1,16 @@
 import { connect } from "react-redux";
 
-import SortButtonsTemplate from "../templates/filtersort/SortButtons";
-import { sort } from "../actions";
+import { capitalize } from "../../utils/string";
+
+import SortDropdownTemplate from "../../templates/filtersort/SortDropdown";
+import { sort } from "../../actions";
 
 const mapStateToProps = (state) => ({
   value: state.list.sortValue,
+  valueLabel: capitalize(state.localization[state.list.sortValue]),
   options: state.list.sortOptions.map((id) => ({
     id,
-    label: state.localization[id],
+    label: capitalize(state.localization[id]),
   })),
   label: state.localization.sort_by_label,
 });
@@ -23,4 +26,4 @@ const mapDispatchToProps = (dispatch) => ({
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(SortButtonsTemplate);
+)(SortDropdownTemplate);
